+++ src/Logging/LogApplicationInformation.cs
@@ -8,9 +8,9 @@ namespace Phoenix.Functionality.Logging;
 /// Contains information about an application.
 /// </summary>
 /// <param name="Name"> The name of the application. </param>
-/// <param name="AlphanumericIdentifier"> A unique alpha-numeric identifier of the application that can be used for example to register the application with a log target or to enrich log events. </param>
 /// <param name="NumericIdentifier"> A unique numeric identifier of the application that can be used for example to register the application with a log target or to enrich log events. </param>
-public record LogApplicationInformation(string Name, string AlphanumericIdentifier, int NumericIdentifier)
+/// <param name="AlphanumericIdentifier"> A unique alpha-numeric identifier of the application that can be used for example to register the application with a log target or to enrich log events. </param>
+public record LogApplicationInformation(string Name, int NumericIdentifier, string AlphanumericIdentifier)
 {
 	/// <summary>
 	/// Starts creating <see cref="LogApplicationInformation"/> via builder pattern.