+++ src/Serilog/⬙/CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ___
 
+## 1.4.0
+
+:calendar: _2023-05-??_
+
+### Added
+
+- The new `ApplicationInformationEnricher` that allows logs to be enriched by a selectable information about an application like **name**, **version**, etc.
+
+### Deprecated
+
+- `ApplicationVersionEnricher` and `ApplicationIdentifierEnricher` (both are replaced by the new `ApplicationInformationEnricher`).
+
+### References
+
+:white_circle: Phoenix.Functionality.Logging **1.0.0**
+___
+
 ## 1.3.0
 
 :calendar: _2023-03-27_